split juseppe to core and cli
