+++ juseppe-cli/src/main/java/ru/lanwen/jenkins/juseppe/files/WatchEventExtension.java
