+++ juseppe-cli/src/main/resources/jetty-logging.properties
