+++ juseppe-core/src/main/java/ru/lanwen/jenkins/juseppe/gen/HPI.java
