+++ juseppe-core/src/main/java/ru/lanwen/jenkins/juseppe/gen/source/PluginSource.java
@@ -1,0 +1,13 @@
+package ru.lanwen.jenkins.juseppe.gen.source;
+
+import ru.lanwen.jenkins.juseppe.beans.Plugin;
+
+import java.util.List;
+
+/**
+ * @author lanwen (Merkushev Kirill)
+ */
+public interface PluginSource {
+
+    List<Plugin> plugins();
+}
