+++ juseppe-core/src/main/java/ru/lanwen/jenkins/juseppe/gen/view/UpdateSiteView.java
@@ -1,0 +1,11 @@
+package ru.lanwen.jenkins.juseppe.gen.view;
+
+/**
+ * @author lanwen (Merkushev Kirill)
+ */
+public interface UpdateSiteView {
+
+    String name();
+
+    String content();
+}
