+++ juseppe-core/src/main/java/ru/lanwen/jenkins/juseppe/util/SignatureGenerator.java
