+++ juseppe-core/src/main/resources/jsonschema/dependency.json
