+++ juseppe-core/src/main/resources/jsonschema/developer.json
