+++ juseppe-core/src/main/resources/jsonschema/plugin.json
