+++ juseppe-core/src/main/resources/jsonschema/release-history.json
