+++ juseppe-core/src/main/resources/jsonschema/signature.json
