+++ juseppe-core/src/main/resources/jsonschema/update-site.json
