+++ juseppe-core/src/test/java/ru/lanwen/jenkins/juseppe/HPICreationTest.java
@@ -1,0 +1,83 @@
+package ru.lanwen.jenkins.juseppe;
+
+import org.junit.ClassRule;
+import org.junit.Test;
+import org.junit.rules.ExternalResource;
+import org.junit.rules.RuleChain;
+import org.junit.rules.TemporaryFolder;
+import ru.lanwen.jenkins.juseppe.beans.UpdateSite;
+import ru.lanwen.jenkins.juseppe.gen.UpdateSiteGen;
+import ru.lanwen.jenkins.juseppe.gen.view.JsonpUpdateSite;
+import ru.lanwen.jenkins.juseppe.props.JuseppeEnvVars;
+import ru.lanwen.jenkins.juseppe.props.Props;
+
+import java.io.File;
+import java.io.IOException;
+import java.net.URI;
+
+import static com.google.common.io.Resources.getResource;
+import static java.lang.System.clearProperty;
+import static java.lang.System.setProperty;
+import static org.hamcrest.CoreMatchers.containsString;
+import static org.hamcrest.Matchers.greaterThan;
+import static org.junit.Assert.assertThat;
+import static ru.lanwen.jenkins.juseppe.gen.UpdateSiteGen.updateSite;
+import static ru.lanwen.jenkins.juseppe.util.JuseppeMatchers.exists;
+
+public class HPICreationTest {
+
+    public static final String BASE_URL_OF_SITE = "http://hudson02.nhncorp.com:9080/update";
+    public static final String PLUGINS_DIR_CLASSPATH = "tmp/plugins";
+
+    public static TemporaryFolder tmp = new TemporaryFolder();
+
+    @ClassRule
+    public static RuleChain setenv = RuleChain.emptyRuleChain().around(tmp).around(new ExternalResource() {
+        @Override
+        protected void before() throws Throwable {
+            folderToSave = tmp.newFolder();
+            System.out.println(folderToSave.getAbsolutePath());
+            clearProperty(JuseppeEnvVars.JuseppeEnvEnum.JUSEPPE_SAVE_TO_DIR.mapping());
+            setProperty(JuseppeEnvVars.JuseppeEnvEnum.JUSEPPE_SAVE_TO_DIR.mapping(), folderToSave.getAbsolutePath());
+        }
+    });
+
+    private static File folderToSave;
+
+    @Test
+    public void shouldSaveJsonWithPluginInfo() throws IOException {
+        File file = tmp.newFile("temp.json");
+
+        updateSite(
+                Props.populated()
+                        .withPluginsDir(getResource(PLUGINS_DIR_CLASSPATH).getFile())
+                        .withBaseurl(URI.create(BASE_URL_OF_SITE))
+                        .withSaveto(file.getParent())
+                        .withUcJsonName(file.getName())
+        ).withDefaults().fill().save();
+
+        assertThat(file, exists());
+        assertThat(file.length(), greaterThan(200l));
+    }
+
+    @Test
+    public void shouldSaveJsonWithReleaseHistory() throws IOException {
+        File file = new File(folderToSave, Props.populated().getReleaseHistoryJsonName());
+
+        UpdateSiteGen.updateSite(Props.populated()
+                .withPluginsDir(getResource(PLUGINS_DIR_CLASSPATH).getFile())).withDefaults().fill().save();
+
+        assertThat(file, exists());
+        assertThat(file.length(), greaterThan(200l));
+    }
+
+    @Test
+    public void shouldContainPlugin() throws IOException {
+        UpdateSite site = UpdateSiteGen.updateSite(Props.populated()
+                .withPluginsDir(getResource(PLUGINS_DIR_CLASSPATH).getFile())).withDefaults().fill().getSite();
+        
+        assertThat(new JsonpUpdateSite(site, "uc.json").content(), containsString("clang-scanbuild-plugin"));
+        assertThat(site.getPlugins().stream().findFirst().get().getUrl(), 
+                containsString(Props.populated().getBaseurl() + "/clang-scanbuild-plugin.hpi"));
+    }
+}
