+++ juseppe-core/src/test/java/ru/lanwen/jenkins/juseppe/props/JuseppeEnvVarsTest.java
