+++ juseppe-core/src/test/java/ru/lanwen/jenkins/juseppe/util/JuseppeMatchers.java
