+++ juseppe-core/src/test/resources/checkstyle-config.xml
