+++ juseppe-core/src/test/resources/log4j.properties
@@ -1,0 +1,7 @@
+# Root logger option
+log4j.rootLogger=TRACE, stdout
+
+# Direct log messages to stdout
+log4j.appender.stdout=org.apache.log4j.ConsoleAppender
+log4j.appender.stdout.layout=org.apache.log4j.PatternLayout
+log4j.appender.stdout.layout.ConversionPattern=%d{yyyy-MM-dd HH:mm:ss} %-5p %m%n
