+++ juseppe-core/src/test/resources/tmp/plugins/clang-scanbuild-plugin.hpi
